float error resolved

+++ components/CreateTicketCard.tsx
@@ -72,7 +72,7 @@ export default function CreateTicketCard({spaces}: { spaces: string[] }) {
             console.log("currentTokenId", currentTokenId)
             const params = {
                 maxSupply: parseInt(form.quantity),
-                mintPrice: parseInt(form.price),
+                mintPrice: parseFloat(form.price),
                 metadataURL: metadataUrl,
                 spaceName: form.space,
             }
